refactor: 더보기 버튼 네이밍 변경 및 contentView 파일 분리
toggleMoreButton -> ExpandFoldButton으로 네이밍 변경
fileprivate ContentView -> ExpandableContentView
